requirements.txt
python-dotenv~=1.0.1
pywinauto~=0.6.8
screeninfo~=0.8.1